Consolidate CI: one workflow for firmware, Docker, releases
- Remove separate ci.yml firmware build and release.yml
- docker.yml now: builds firmware, uploads artifact, creates GitHub release on tags, builds Docker image
- ci.yml now only builds PC simulator (fast feedback)
- Fix format string warning in ui_jpeg.c for ESP-IDF 5.4

+++ common/ui_jpeg.c
@@ -147,7 +147,7 @@ bool ui_jpeg_decode_to_lvgl(const uint8_t *jpeg_data,
     free(io);
     free(info);
 
-    ESP_LOGI(TAG, "Decoded JPEG to %dx%d RGB565 (%zu bytes)", w, h, out_img->dsc.data_size);
+    ESP_LOGI(TAG, "Decoded JPEG to %dx%d RGB565 (%lu bytes)", w, h, (unsigned long)out_img->dsc.data_size);
     return true;
 }
